test modified and all passed
