+++ frontend/progetto-risiko-risikiamo-frontend/package.json
@@ -10,11 +10,13 @@
   },
   "dependencies": {
     "ammap3": "^3.21.15",
-    "core-js": "^3.8.3",
-    "vue": "^3.2.13",
     "axios": "^0.27.2",
+    "core-js": "^3.8.3",
+    "typescript-require": "^0.3.0",
+    "vue": "^3.2.13",
+    "vuex": "4.1.0",
     "vue-router": "^4.0.13",
-    "typescript-require": "^0.3.0"
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
