+++ frontend/progetto-risiko-risikiamo-frontend/src/components/NestedRoute.vue
@@ -1,0 +1,13 @@
+<template>
+    <div>
+        NestedRoute
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'NestedRoute',
+    components: {
+    }
+}
+</script>
