+++ frontend/progetto-risiko-risikiamo-frontend/src/routers.js
@@ -1,11 +1,17 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeVue from './components/HomeComponent.vue'
+import NestedRoute from "./components/NestedRoute.vue"
 
 const routes = [
     {
         name: "home",
         component: HomeVue,
         path: "/"
+    },
+    {
+        name: "nested-route",
+        component: NestedRoute,
+        path: "/nested-route"
     }
 ];
